requirements.txt
django==3.1
djangorestframework==3.12.4
pillow==7.2.0
django-storages==1.11.1
boto3==1.18.0
django-environ==0.4.5
pytest-django==4.4.0
mixer==7.1.2
djangorestframework-api-key==2.*
